Extract P2PTxInvStore into test framework
